Fix tests

+++ src/main/java/Inventory.java
@@ -3,8 +3,6 @@ import java.util.LinkedList;
 import java.util.List;
 
 public class Inventory {
-
-
     private List phones;
 
     Inventory() {
@@ -12,13 +10,13 @@ public class Inventory {
     }
 
     void addPhone(String model, String serialNumber, Type type,
-                  float price, boolean used) {
+                  double price, boolean used) {
         Phone phone = new Phone(model, serialNumber, type, price, used);
         phones.add(phone);
     }
 
     Phone getPhone(String serialNumber) {
-        for (Iterator phoneCheck = phones.iterator(); phoneCheck.hasNext();) {
+        for (Iterator phoneCheck = phones.iterator(); phoneCheck.hasNext(); ) {
             Phone phone = (Phone) phoneCheck.next();
             if (phone.getSerialNumber().equals(serialNumber)) {
                 return phone;
@@ -29,7 +27,7 @@ public class Inventory {
 
     public List search(Phone searchPhone) {
         List matchingPhones = new LinkedList();
-        for (Iterator phoneCheck = phones.iterator(); phoneCheck.hasNext();) {
+        for (Iterator phoneCheck = phones.iterator(); phoneCheck.hasNext(); ) {
             Phone phone = (Phone) phoneCheck.next();
             if (phone.getSpec().sameAs(searchPhone.getSpec())) {
                 continue;
