+++ src/main/java/Phone.java
@@ -4,7 +4,7 @@ class Phone {
     private String serialNumber;
 
     Phone(String model, String serialNumber, Type type,
-          float price, boolean used) {
+          double price, boolean used) {
         this.spec = new PhoneSpec(type, model, used);
         this.serialNumber = serialNumber;
         this.price = price;
